Adaption to Gnoga 2.1a and reduction of flickering.
The GUI BB depends now on Gnoga 2.1a.

Back to the browser version for the GUI, instead of native Gtk window - GtkWindow in Gnoga requires libwebkitgtk-3.0 and other libraries that are not available on current Debian, which makes it.

Change of approach in the body of BB.GUI.Controller to reduce flickering.